# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuildGraph AI MVP API
  version: 0.2.0
  description: BuildGraph 3주 MVP 공통 API 계약. 상세 기준은 docs/API_CONTRACT.md를 따른다.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/users:
    post:
      summary: 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: 생성된 사용자
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
  /api/auth/login:
    post:
      summary: 이메일 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT와 사용자 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/price-alerts:
    get:
      summary: 목표가 알림 목록
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 목표가 알림 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlertPage"
    post:
      summary: 목표가 알림 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceAlertCreateRequest"
      responses:
        "201":
          description: 생성된 알림
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlertDto"
  /api/admin/price-jobs:
    get:
      summary: 관리자 가격 수집 작업 목록
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 가격 수집 작업 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceJobPage"
  /api/admin/price-jobs/run:
    post:
      summary: 가격 수집 작업 실행
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceJobRunRequest"
      responses:
        "201":
          description: 생성된 가격 수집 작업
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceJobDto"
...
  /api/agent/sessions:
    post:
      summary: Agent 세션 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentSessionCreateRequest"
      responses:
        "201":
          description: 생성된 Agent 세션
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSessionDto"
  /api/agent/sessions/{id}/run:
    post:
      summary: Agent 세션 실행 시작
      parameters:
        - $ref: "#/components/parameters/PublicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "200":
          description: RUNNING 상태로 전이된 Agent 세션
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSessionDto"
  /api/agent/sessions/{id}:
    get:
      summary: Agent 세션 상세
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: Agent 세션 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSessionDto"
...
  /api/agent-logs/upload:
    post:
      summary: PC Agent JSONL 로그 업로드
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/AgentLogUploadRequest"
      responses:
        "201":
          description: 업로드된 로그 메타데이터
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentLogUploadDto"
  /api/agent-logs/{id}:
    get:
      summary: Agent 로그 업로드 상세
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: 로그 업로드 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentLogUploadDto"
  /api/as-tickets:
    post:
      summary: AS 티켓 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AsTicketCreateRequest"
      responses:
        "201":
          description: 생성된 AS 티켓
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsTicketDto"
  /api/as-tickets/{id}:
    get:
      summary: AS 티켓 상세
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: AS 티켓 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsTicketDto"
...
  /api/admin/agent-sessions:
    get:
      summary: 관리자 Agent 세션 목록
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Agent 세션 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSessionPage"
  /api/admin/agent-sessions/{id}:
    get:
      summary: 관리자 Agent 세션 상세
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: Agent 세션 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAgentSessionDto"
  /api/admin/tool-invocations:
    get:
      summary: 관리자 Tool 호출 목록
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Tool 호출 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolInvocationPage"
  /api/admin/tool-invocations/{id}:
    get:
      summary: 관리자 Tool 호출 상세
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: Tool 호출 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolInvocationDto"
  /api/admin/rag-evidence/{id}:
    get:
      summary: 관리자 RAG 근거 상세
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: 관리자 RAG 근거 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRagEvidenceDto"
  /api/admin/as-tickets:
    get:
      summary: 관리자 AS 티켓 목록
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: AS 티켓 목록
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsTicketPage"
  /api/admin/as-tickets/{id}:
    get:
      summary: 관리자 AS 티켓 상세
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: 관리자 AS 티켓 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsTicketDto"
    patch:
      summary: 관리자 AS 티켓 수정
      parameters:
        - $ref: "#/components/parameters/PublicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAsTicketUpdateRequest"
      responses:
        "200":
          description: 수정된 AS 티켓
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsTicketDto"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PublicId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      in: query
      name: size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    Role:
      type: string
      enum: [USER, ADMIN]
...
    ToolStatus:
      type: string
      enum: [PASS, WARN, FAIL]
    Confidence:
      type: string
      enum: [LOW, MEDIUM, HIGH]
...
    AgentSessionStatus:
      type: string
      enum: [QUEUED, RUNNING, RAG_SEARCHED, TOOLS_CALLED, SUMMARY_READY, FALLBACK_READY, SUCCEEDED, FAILED, CANCELLED]
    AsTicketStatus:
      type: string
      enum: [OPEN, ASSIGNED, IN_PROGRESS, RESOLVED, CLOSED, CANCELLED]
    LogUploadStatus:
      type: string
      enum: [UPLOADED, PROCESSING, FAILED]
    PriceAlertStatus:
      type: string
      enum: [ACTIVE, TRIGGERED, DISABLED]
    PriceJobStatus:
      type: string
      enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
    UserDto:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
    SignupRequest:
      type: object
      required: [email, password, name, termsAccepted, marketingAccepted]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        termsAccepted:
          type: boolean
        marketingAccepted:
          type: boolean
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserDto"
...
    PriceAlertCreateRequest:
      type: object
      required: [partId, targetPrice]
      properties:
        partId:
          type: string
          format: uuid
        targetPrice:
          type: integer
    PriceAlertDto:
      type: object
      required: [partId, partName, targetPrice, currentPrice, status]
      properties:
        partId:
          type: string
          format: uuid
        partName:
          type: string
        targetPrice:
          type: integer
        currentPrice:
          type: integer
        status:
          $ref: "#/components/schemas/PriceAlertStatus"
        createdAt:
          type: string
          format: date-time
    PriceJobRunRequest:
      type: object
      properties:
        source:
          type: string
          example: manual
    PriceJobDto:
      type: object
      required: [id, status, requestedBy]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/PriceJobStatus"
        requestedBy:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        errorSummary:
          type: string
        createdAt:
          type: string
          format: date-time
...
    AgentSessionCreateRequest:
      type: object
      properties:
        requirementId:
          type: string
          format: uuid
          nullable: true
        buildId:
          type: string
          format: uuid
          nullable: true
        asTicketId:
          type: string
          format: uuid
          nullable: true
      description: requirementId, buildId, asTicketId 중 정확히 하나만 non-null이어야 한다.
    AgentSessionDto:
      type: object
      required: [id, status, stateTimeline]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AgentSessionStatus"
        summary:
          type: string
          nullable: true
        stateTimeline:
          type: array
          items:
            $ref: "#/components/schemas/StateTimelineItemDto"
        toolInvocationIds:
          type: array
          items:
            type: string
            format: uuid
        evidenceIds:
          type: array
          items:
            type: string
            format: uuid
    AdminAgentSessionDto:
      allOf:
        - $ref: "#/components/schemas/AgentSessionDto"
        - type: object
          properties:
            toolInvocations:
              type: array
              items:
                $ref: "#/components/schemas/ToolInvocationDto"
    StateTimelineItemDto:
      type: object
      required: [to, actor, at]
      properties:
        from:
          type: string
          nullable: true
        to:
          $ref: "#/components/schemas/AgentSessionStatus"
        actor:
          type: string
          enum: [USER, ADMIN, SYSTEM]
        at:
          type: string
          format: date-time
        reason:
          type: string
    RagEvidenceDto:
      type: object
      required: [id, sourceId, summary]
      properties:
        id:
          type: string
          format: uuid
        sourceId:
          type: string
        summary:
          type: string
        score:
          type: number
        metadata:
          type: object
          additionalProperties: true
    AdminRagEvidenceDto:
      allOf:
        - $ref: "#/components/schemas/RagEvidenceDto"
        - type: object
          properties:
            agentSessionId:
              type: string
              format: uuid
            chunkText:
              type: string
    ToolInvocationDto:
      type: object
      required: [id, agentSessionId, toolName, status, confidence, summary]
      properties:
        id:
          type: string
          format: uuid
        agentSessionId:
          type: string
          format: uuid
        toolName:
          type: string
          enum: [compatibility, power, size, performance, price]
        status:
          $ref: "#/components/schemas/ToolStatus"
        confidence:
          $ref: "#/components/schemas/Confidence"
        summary:
          type: string
        requestPayload:
          type: object
          additionalProperties: true
        resultPayload:
          type: object
          additionalProperties: true
        latencyMs:
          type: integer
        createdAt:
          type: string
          format: date-time
    AgentLogUploadRequest:
      type: object
      required: [file, rangeMinutes, consentAccepted]
      properties:
        file:
          type: string
          format: binary
        rangeMinutes:
          type: integer
        consentAccepted:
          type: boolean
    AgentLogUploadDto:
      type: object
      required: [id, status, fileName, rangeMinutes, deleteAfter]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/LogUploadStatus"
        fileName:
          type: string
        fileSize:
          type: integer
        rangeMinutes:
          type: integer
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
        deleteAfter:
          type: string
          format: date-time
    AsTicketCreateRequest:
      type: object
      required: [symptom]
      properties:
        logUploadId:
          type: string
          format: uuid
        symptom:
          type: string
    AsTicketDto:
      type: object
      required: [id, status, symptom, causeCandidates, upgradeCandidates]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AsTicketStatus"
        symptom:
          type: string
        logUploadId:
          type: string
          format: uuid
        assignedAdminId:
          type: string
          format: uuid
        causeCandidates:
          type: array
          items:
            type: object
            additionalProperties: true
        upgradeCandidates:
          type: array
          items:
            type: object
            additionalProperties: true
        adminNote:
          type: string
        resolvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminAsTicketUpdateRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/AsTicketStatus"
        assignedAdminId:
          type: string
          format: uuid
        adminNote:
          type: string
...
    PriceAlertPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PriceAlertDto"
    PriceJobPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PriceJobDto"
    AgentSessionPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AgentSessionDto"
    ToolInvocationPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ToolInvocationDto"
...
    AsTicketPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AsTicketDto"
...
    PageMeta:
      type: object
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items: {}
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer